chore: refactor modifier
