+++ ouds_core/lib/components/navigation/internal/ouds_navigation_bar_modifier.dart
@@ -11,8 +11,6 @@
  * //
  */
 
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
 import 'package:ouds_core/components/navigation/internal/ouds_navigation_bar_state.dart';
 import 'package:ouds_theme_contract/ouds_theme.dart';
@@ -22,38 +20,6 @@ class OudsNavigationBarStatusModifier {
   final BuildContext context;
 
   OudsNavigationBarStatusModifier(this.context);
-
-  /// Returns the background color of the navigation bar depending on translucency and theme.
-  Color getBackgroundColor(bool isTranslucent) {
-    final barTheme = OudsTheme.of(context).componentsTokens(context).bar;
-    final isDark = ThemeUtils.isDarkTheme(context);
-    if (isTranslucent) {
-      return isDark ? barTheme.colorBgTranslucentDark.withValues(alpha: 0.5) : barTheme.colorBgTranslucentLight.withValues(alpha: 0.5);
-    } else {
-      return barTheme.colorBgOpaque.withValues(alpha: 0.5);
-    }
-  }
-
-  /// Returns a Border applied to the top edge of the navigation bar.
-  /// The BorderSide's color and width are derived from the current theme's border tokens.
-  Border getBorderNavigationBar() {
-    final theme = OudsTheme.of(context);
-    return Border(
-      top: BorderSide(
-        color: theme.colorScheme(context).borderMinimal,
-        width: theme.borderTokens.widthDefault,
-      ),
-    );
-  }
-
-  /// Returns an ImageFilter that applies a background blur for the navigation bar.
-  ImageFilter getBlurNavigationBar() {
-    final barTheme = OudsTheme.of(context).componentsTokens(context).bar;
-    return ImageFilter.blur(
-      sigmaX: barTheme.effectBgBlur.toDouble(),
-      sigmaY: barTheme.effectBgBlur.toDouble(),
-    );
-  }
 
   /// Returns the text color of a navigation item based on its state and selection.
   Color getTextIconItemColor(OudsNavigationBarControlState state, [bool isSelected = false]) {
